# Allow FrameReader buffer to be uninitialized