Summary:
Yields another ~10% improvement in the pingpong channel benchmark:

Before:
```
Results:
Average latency: 0.58219533 ms
Min latency: 0.342 ms
Max latency: 11.044 ms
Total iterations: 100000
Total time: 58 seconds
Bytes sent: 100000000000 bytes
Bytes received: 100000000000 bytes
Total bytes transferred: 200000000000 bytes
Bandwidth: 3434714661.0795307 bytes/sec (26204.79325164437 Mbps)
```
After:
```
Results:
Average latency: 0.52789287 ms
Min latency: 0.286 ms
Max latency: 12.973 ms
Total iterations: 100000
Total time: 52 seconds
Bytes sent: 100000000000 bytes
Bytes received: 100000000000 bytes
Total bytes transferred: 200000000000 bytes
Bandwidth: 3788022273.5709686 bytes/sec (28900.31641823554 Mbps)
```

Differential Revision: D83482464